Move chat suggestions into their own file
